feat(widget): move PageIndicator widget to common folder
